Moved signup form models into their own app
