+++ tbx/services/templates/blocks/services/sign_up_form_page_block.html
@@ -3,7 +3,7 @@
 <div class="services-section">
 	<div class="services-wrapper">
 		{% with page_id=value.page.pk|stringformat:"s" %}
-		    {% include 'torchbox/includes/sign_up_form_page_form.html' with page=value.page legend=value.page.call_to_action_text form=form form_id='sign-up-form-'|add:page_id %}
+		    {% include 'sign_up_form/includes/sign_up_form_page_form.html' with page=value.page legend=value.page.call_to_action_text form=form form_id='sign-up-form-'|add:page_id %}
 		{% endwith %}
 	</div>
 </div>
