+++ tbx/sign_up_form/templates/sign_up_form/sign_up_form_page.html
@@ -12,7 +12,7 @@
     <div class="container">
         {{ page.intro|richtext }}
         <div class="form sign-up-form-wrapper top">
-            {% include "torchbox/includes/sign_up_form_page_form.html" with legend=page.call_to_action_text %}
+            {% include "sign_up_form/includes/sign_up_form_page_form.html" with legend=page.call_to_action_text %}
         </div> <!-- /sign-up-form -->
     </div>
 
@@ -58,7 +58,7 @@
     <div class="cta">
         <div class="container">
             <div class="form sign-up-form-wrapper">
-                {% include "torchbox/includes/sign_up_form_page_form.html" with legend=page.call_to_action_text %}
+                {% include "sign_up_form/includes/sign_up_form_page_form.html" with legend=page.call_to_action_text %}
             </div> <!-- /sign-up-form -->
 
         </div>
